Updated Bathroom.java's getLongtiude() and getLatitude() methods to return double in order to match Coordinates.java's longitude and latitude types.
Fixed ServerInterfacer.java's bathroomSearch() method's floor parameter
to be a String to match Bathroom.java's floor field.

+++ src/edu/washington/cs/lavatorylocator/Bathroom.java
@@ -86,7 +86,7 @@ public class Bathroom {
      * 
      * @return the longitude of this bathroom
      */
-    public float getLongitude() {
+    public double getLongitude() {
         return location.getLongitude();
     }
 
@@ -95,7 +95,7 @@ public class Bathroom {
      * 
      * @return the latitude of this bathroom
      */
-    public float getLatitude() {
+    public double getLatitude() {
         return location.getLatitude();
     }
